+++ app/Config.example.php
@@ -140,6 +140,7 @@ Config::set('modules', array(
 Config::set('classAliases', array(
     'Config'        => '\Core\Config',
     'Errors'        => '\Core\Error',
+    'Request'       => '\Core\Request',
     'Response'      => '\Core\Response',
     'Redirect'      => '\Core\Redirect',
     'Mail'          => '\Helpers\Mailer',
@@ -160,7 +161,6 @@ Config::set('classAliases', array(
     'Paginator'     => '\Helpers\Paginator',
     'Password'      => '\Helpers\Password',
     'RainCaptcha'   => '\Helpers\RainCaptcha',
-    'Request'       => '\Helpers\Request',
     'ReservedWords' => '\Helpers\ReservedWords',
     'Session'       => '\Helpers\Session',
     'SimpleCurl'    => '\Helpers\SimpleCurl',
