+++ system/Core/Request.php
@@ -5,7 +5,7 @@
  * @version 3.0
  */
 
-namespace Helpers;
+namespace Core;
 
 /**
  * Contains the Request information and provides methods to fetch the HTTP Request body.
